+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## [0.1.0] 2025-11
 
+### Added 
 - Add initial Makie theme for DIE
-- 
+- Helper plotting functions `tslines` and `tslines!` for plotting DIE-styled time series. 
+- New function `annotate_lines!`: annotates all (or just the last) lines in the given axis. 
+- New functions: `dateticks!` sets new x-ticks with specified dates and formats, `dateszoom!` zooms the x-axis over the specified dates.